code/requirements.txt
gymnasium
numpy
matplotlib
pyyaml
pandas
stable-baselines3==2.0.0
torch>=1.12.0
tensorboard
